backend/requirements.txt
fastapi==0.120.4
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.22
psycopg2-binary==2.9.9

pydantic==2.6.1
pydantic-settings==2.2.1

python-dotenv==1.0.0
requests==2.32.1

alembic==1.12.1
python-multipart
jinja2
